// src/app/api/models/group-member-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * DTO for group member information
 */
export interface GroupMemberDto {

  /**
   * Full name of the user
   */
  fullName?: string;

  /**
   * Group ID
   */
  groupId?: string;

  /**
   * Date and time when the user joined the group
   */
  joinedAt?: string;

  /**
   * Member ID
   */
  memberId?: string;

  /**
   * Represents the role of a user within a group
   */
  role?: 'member' | 'admin';

  /**
   * User ID
   */
  userId?: string;

  /**
   * User name
   */
  userName?: string;
}